refactor : separate `fetchNotification` function into shared/util
