+++ src/shared/utils/index.ts
@@ -4,3 +4,4 @@ export * from './getCookie';
 export * from './getFormattedPhone';
 export * from './fetchLoginStatus';
 export * from './getPath';
+export * from './fetchNotification';
